refactor(stella-core): one turn loop, not three — extract Engine::drive
`Engine::run_step` was already the only implementation of a step. The loop
around it was written three times: the fresh turn in
`run_turn_with_sender`, the restored turn in `driver/resume.rs`, and the
served turn in `stella-serve`'s session thread. The three differed only in
how a `TurnState` is minted and how the outcome is settled back, and the
served copy had already forgotten an obligation — an `EngineConfig::turn_halt`
armed by an embedder through the public `SessionSpec::config` was never
consulted.

`Engine::drive` is that loop, once: turn framing (the opening
`Stage(Execute)` and the `agent.turn.started`/`completed` bus pair), the
step-cap gate, `run_step`, `persist_checkpoint` on `Continue`, the
`turn_halt` consultation at the committed boundary, and `discard_checkpoint`
on every terminal path.

The framing moves too, which #2452 had scoped out. It is the same defect in
the same shape: `stella-serve` already calls `engine.with_bus(bus)` and
`with_turn_instance` carries the bus through, so the engine's `emit_lifecycle`
and serve's hand-rolled `bus.emit_named` were the same bus emitting the same
shapers. Folding it in also retires the `bus` parameter threaded through
`drive_turn` and `drive_goal`.

`drive_restored_turn` becomes `self.drive(...)` and nothing else, so it goes;
its two callers (`stella-cli`'s daemon resume, `stella-pipeline`'s resumed
execute stage) call `drive` directly. Its module docs and both its tests move
to `driver/drive.rs` unchanged.

Refs #2452
